refactor : clubInfo 패키지 수정
