+++ src/main/java/ddingdong/ddingdongBE/domain/feed/service/dto/query/ClubInfo.java
@@ -1,4 +1,4 @@
-package ddingdong.ddingdongBE.domain.feed.vo;
+package ddingdong.ddingdongBE.domain.feed.service.dto.query;
 
 import io.swagger.v3.oas.annotations.media.Schema;
 import lombok.Builder;
